stoneProviderをquiz.dartから切り離し、ロジック部分をstone_provider.dartで管理
